+++ src/Button/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Button.module.css';
+import styles from './index.module.css';
 
 class Button extends React.Component {
   constructor(props) {
