+++ src/Input/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Input.module.css';
+import styles from './index.module.css';
 class Input extends React.Component {
   constructor(props) {
     super(props);
@@ -15,10 +15,18 @@ class Input extends React.Component {
       placeholder = inputLabel,
     } = this.props;
     const errorMessage = state.errors[name];
+    const value = state[name];
+    const numOfSymbols = value ? value.length : 0;
+    console.log(value);
+    console.log(numOfSymbols);
+
     if (isTextArea) {
       return (
         <div className={styles.formInput}>
-          <label htmlFor={name}>{inputLabel}</label>
+          <label htmlFor={name}>
+            <div>{inputLabel}</div>
+            <div className={styles.symbolsLimit}>{`${numOfSymbols}/600`}</div>
+          </label>
           <textarea
             id={name}
             name={name}
@@ -26,7 +34,7 @@ class Input extends React.Component {
             rows="7"
             placeholder={placeholder}
             onChange={onChange}
-            value={state[name]}
+            value={value}
           />
           <div className={errorMessage ? styles.errorMessage : styles.invisible}>
             {errorMessage}
