+++ src/Input/index.module.css
@@ -15,10 +15,11 @@ input {
   border-radius: 5px;
 }
 label {
+  display: flex;
+  column-gap: 10px;
   font-weight: 800;
   color: rgb(96, 94, 94);
   font-size: 22px;
-  display: block;
   width: 100%;
 }
 
@@ -43,6 +44,12 @@ label {
   width: 100%;
   min-height: 100px;
 }
+.symbolsLimit {
+  font-size: 14px;
+  padding-top: 5px;
+  font-weight: 400;
+  align-self: center;
+}
 
 .errorMessage {
   font-size: 14px;
@@ -53,4 +60,8 @@ label {
 .invisible {
   color: transparent;
   min-height: 30px;
+}
+
+::placeholder {
+  color: rgb(151, 150, 150);
 }
